Move sgn0 to bellpepper-emulated
Also undo the pub(crate) changes
